data/gathering_data/gtfs/requirements.txt
tqdm
requests
tenacity
urllib3
python-dotenv